requirements.txt
numpy
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
openai
tiktoken
-e .